refactor: reorganize files and collapse store
