+++ internal/api/scim/server_test.go
@@ -549,11 +549,3 @@ func listed[T any](t *testing.T, w *httptest.ResponseRecorder) protocol.ListResp
 	require.NoError(t, json.Unmarshal(w.Body.Bytes(), &body))
 	return body
 }
-
-func userNamesOf(users []*core.User) []string {
-	names := make([]string, 0, len(users))
-	for _, user := range users {
-		names = append(names, user.UserName)
-	}
-	return names
-}
