+++ internal/api/scim/user_repository.go
@@ -12,22 +12,13 @@ import (
 
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgerrcode"
-
 	"github.com/supabase/auth/internal/api/scim/core"
 	"github.com/supabase/auth/internal/api/scim/protocol"
 	"github.com/supabase/auth/internal/storage"
 )
 
-// scimUsersTable is unqualified so that it resolves through the connection's
-// search_path, as every table named in internal/models does. Which schema that
-// is comes from DB_NAMESPACE.
 const scimUsersTable = "scim_users"
-
-// scimUserWhere is the predicate every query in this repository carries. It is
-// written once so that no query can be assembled without it: the tenant scopes
-// the row, and a soft-deleted row is gone.
 const scimUserWhere = " WHERE sso_provider_id = ? AND deleted_at IS NULL"
-
 const scimUserColumns = "id, resource, created_at, updated_at"
 
 // userSortColumns is what a client may order Users by, mapped to the SQL that
@@ -46,36 +37,23 @@ var userSortColumns = map[string]string{
 	"meta.lastmodified": "updated_at",
 }
 
-type userStore struct {
+type userRepository struct {
 	db       *storage.Connection
 	usersURL string
+	tenant   string
 }
 
-// NewUserStore serves Users out of Postgres. externalURL is where this
-// deployment answers, which is what makes a stored row addressable as
-// meta.location.
-func NewUserStore(db *storage.Connection, externalURL string) Store[*core.User] {
-	return &userStore{
-		db:       db,
-		usersURL: core.KindUser.Location(core.Join(externalURL, BasePath)),
-	}
-}
-
-func (s *userStore) For(tenant string) Repository[*core.User] {
-	return &userRepository{store: s, tenant: tenant}
-}
-
-// userRepository is one tenant's Users. The tenant is a field rather than a
-// parameter so that every query it builds carries scimUserWhere.
-type userRepository struct {
-	store  *userStore
-	tenant string
+type scimUserRow struct {
+	ID        string    `db:"id"`
+	Resource  []byte    `db:"resource"`
+	CreatedAt time.Time `db:"created_at"`
+	UpdatedAt time.Time `db:"updated_at"`
 }
 
 func (r *userRepository) Get(ctx context.Context, id string) (*core.User, error) {
 	var rows []scimUserRow
 
-	err := r.store.db.WithContext(ctx).RawQuery(
+	err := r.db.WithContext(ctx).RawQuery(
 		"SELECT "+scimUserColumns+" FROM "+scimUsersTable+scimUserWhere+" AND id = ?",
 		r.tenant, id,
 	).All(&rows)
@@ -86,7 +64,7 @@ func (r *userRepository) Get(ctx context.Context, id string) (*core.User, error)
 	if len(rows) == 0 {
 		return nil, ErrNotFound
 	}
-	return r.store.user(rows[0])
+	return r.user(rows[0])
 }
 
 func (r *userRepository) List(ctx context.Context, query *protocol.SearchRequest) ([]*core.User, int, error) {
@@ -100,7 +78,7 @@ func (r *userRepository) List(ctx context.Context, query *protocol.SearchRequest
 		return nil, 0, err
 	}
 
-	db := r.store.db.WithContext(ctx)
+	db := r.db.WithContext(ctx)
 
 	var total int
 	if err := db.RawQuery(
@@ -128,7 +106,7 @@ func (r *userRepository) List(ctx context.Context, query *protocol.SearchRequest
 
 	users := make([]*core.User, 0, len(rows))
 	for _, row := range rows {
-		user, err := r.store.user(row)
+		user, err := r.user(row)
 		if err != nil {
 			return nil, 0, err
 		}
@@ -144,14 +122,14 @@ func (r *userRepository) Create(ctx context.Context, user *core.User) (*core.Use
 	}
 
 	var rows []scimUserRow
-	if err := r.store.db.WithContext(ctx).RawQuery(
+	if err := r.db.WithContext(ctx).RawQuery(
 		"INSERT INTO "+scimUsersTable+" (sso_provider_id, resource) VALUES (?, ?)"+
 			" RETURNING "+scimUserColumns,
 		r.tenant, document,
 	).All(&rows); err != nil {
 		return nil, r.writeError("creating", err)
 	}
-	return r.store.user(rows[0])
+	return r.user(rows[0])
 }
 
 func (r *userRepository) Replace(ctx context.Context, id string, user *core.User) (*core.User, error) {
@@ -161,7 +139,7 @@ func (r *userRepository) Replace(ctx context.Context, id string, user *core.User
 	}
 
 	var rows []scimUserRow
-	if err := r.store.db.WithContext(ctx).RawQuery(
+	if err := r.db.WithContext(ctx).RawQuery(
 		"UPDATE "+scimUsersTable+" SET resource = ?, updated_at = now()"+scimUserWhere+
 			" AND id = ? RETURNING "+scimUserColumns,
 		document, r.tenant, id,
@@ -171,13 +149,9 @@ func (r *userRepository) Replace(ctx context.Context, id string, user *core.User
 	if len(rows) == 0 {
 		return nil, ErrNotFound
 	}
-	return r.store.user(rows[0])
+	return r.user(rows[0])
 }
 
-// Patch reads the resource, applies the operations, and writes it back. It is
-// two statements rather than one because SCIM patch semantics live in the
-// resource document, not in SQL; a single provisioner serialises its own writes,
-// so the read and the write do not race in practice.
 func (r *userRepository) Patch(ctx context.Context, id string, patch *protocol.PatchOp) (*core.User, error) {
 	current, err := r.Get(ctx, id)
 	if err != nil {
@@ -191,13 +165,9 @@ func (r *userRepository) Patch(ctx context.Context, id string, patch *protocol.P
 	return r.Replace(ctx, id, patched)
 }
 
-// Delete is a soft delete: the row stays, its deleted_at is set, and every read
-// filters it out through scimUserWhere. That is what a SCIM DELETE and Okta's
-// deactivate both want, and it lets a userName be reused once its holder is
-// gone.
 func (r *userRepository) Delete(ctx context.Context, id string) error {
 	var ids []string
-	if err := r.store.db.WithContext(ctx).RawQuery(
+	if err := r.db.WithContext(ctx).RawQuery(
 		"UPDATE "+scimUsersTable+" SET deleted_at = now()"+scimUserWhere+" AND id = ? RETURNING id",
 		r.tenant, id,
 	).All(&ids); err != nil {
@@ -209,9 +179,6 @@ func (r *userRepository) Delete(ctx context.Context, id string) error {
 	return nil
 }
 
-// userDocument is the resource as it is stored: the client-supplied attributes
-// only. id and meta are the server's, kept in real columns and reconstructed on
-// read, so they are cleared from the document rather than trusted from it.
 func userDocument(user *core.User) ([]byte, error) {
 	stored := *user
 	stored.ID = ""
@@ -259,6 +226,28 @@ func (r *userRepository) where(query *protocol.SearchRequest) (string, []any, er
 	return where, args, nil
 }
 
+func (r *userRepository) user(row scimUserRow) (*core.User, error) {
+	user := new(core.User)
+	if err := json.Unmarshal(row.Resource, user); err != nil {
+		return nil, fmt.Errorf("scim: decoding stored user %s: %w", row.ID, err)
+	}
+
+	user.ID = row.ID
+	user.Meta = core.Meta{
+		ResourceType: core.KindUser.Name,
+		Created:      row.CreatedAt.UTC(),
+		LastModified: row.UpdatedAt.UTC(),
+		Location:     core.Join(r.usersURL, row.ID),
+	}
+
+	// A stored User is a User whether or not the document says so, and
+	// Section 3.1 does not make "schemas" optional in a response.
+	if len(user.Schemas) == 0 {
+		user.Schemas = []core.SchemaURI{core.SchemaUser}
+	}
+	return user, nil
+}
+
 // userOrderBy is the total order the query asks for. Every sort breaks its ties
 // on id, because StartIndex and Count are a window and a window over a partial
 // order silently skips and repeats rows between pages. Descending reverses the
@@ -287,36 +276,4 @@ func userOrderBy(query *protocol.SearchRequest) (string, error) {
 		return column + direction, nil
 	}
 	return column + direction + ", id" + direction, nil
-}
-
-type scimUserRow struct {
-	ID        string    `db:"id"`
-	Resource  []byte    `db:"resource"`
-	CreatedAt time.Time `db:"created_at"`
-	UpdatedAt time.Time `db:"updated_at"`
-}
-
-// user rebuilds a resource from its row. The document holds what a client
-// supplied; id and meta are this server's to assert, per RFC 7643,
-// Section 3.1, so they come from the columns rather than the document.
-func (s *userStore) user(row scimUserRow) (*core.User, error) {
-	user := new(core.User)
-	if err := json.Unmarshal(row.Resource, user); err != nil {
-		return nil, fmt.Errorf("scim: decoding stored user %s: %w", row.ID, err)
-	}
-
-	user.ID = row.ID
-	user.Meta = core.Meta{
-		ResourceType: core.KindUser.Name,
-		Created:      row.CreatedAt.UTC(),
-		LastModified: row.UpdatedAt.UTC(),
-		Location:     core.Join(s.usersURL, row.ID),
-	}
-
-	// A stored User is a User whether or not the document says so, and
-	// Section 3.1 does not make "schemas" optional in a response.
-	if len(user.Schemas) == 0 {
-		user.Schemas = []core.SchemaURI{core.SchemaUser}
-	}
-	return user, nil
 }
