--- internal/api/scim/users_test.go
@@ -1,1 +1,0 @@
-package scim
